1일차 수정
