+++ src/App.js
@@ -1,20 +1,19 @@
-import Home from './page/Home';
-import Write from './page/Reply';
-import Reply from './page/Reply';
+import Home from './pages/Home';
 import Write from './pages/Write';
+import Reply from './pages/Reply.jsx';
 import styled from 'styled-components';
 
 import { Routes, Route } from 'react-router-dom';
 
 function App() {
   return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/Write" element={<Write />} />
-      <Route path="/Reply" element={<Reply />} />
-        <Home />
-        <Write />
-    </Routes>
+    <div>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/Write" element={<Write />} />
+        <Route path="/Reply" element={<Reply />} />
+      </Routes>
+    </div>
   );
 }
