--- src/components/Client.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import LikeButton from '/LikeButton';
-
-ReactDOM.render(<LikeButton />, document.getElementById('root'));
