+++ src/components/homeIcon/DeleteIcon.jsx
