+++ src/components/write/Button.jsx
@@ -1,15 +1,14 @@
-import styled from "styled-components"
+import styled from 'styled-components';
 
 const Button = () => {
-  
   return (
     <>
       <StButton>작성 완료</StButton>
     </>
-  )
-}
+  );
+};
 
-export default Button
+export default Button;
 
 const StButton = styled.button`
   width: 25%;
@@ -18,12 +17,12 @@ const StButton = styled.button`
   font-size: 1.2em;
   font-weight: bold;
 
-  border-radius:10px;
-  border:none;
-  
-  background-color:#fff;
+  border-radius: 10px;
+  border: none;
+
+  background-color: #fff;
 
   box-shadow: 5px 5px 5px gray;
 
   cursor: pointer;
-`
+`;
