+++ src/components/write/Form.jsx
@@ -1,0 +1,11 @@
+import InputBox from './InputBox';
+
+const Form = () => {
+  return (
+    <div style={{ height: '100vh' }}>
+      <InputBox />
+    </div>
+  );
+};
+
+export default Form;
