--- src/components/writeComponents/Form.jsx
@@ -1,12 +1,0 @@
-import InputBox from "./InputBox"
-
-const Form = () => {
-
-  return (
-    <div style={{height: '100vh'}}>
-      <InputBox />
-    </div>
-  )
-}
-
-export default Form
