--- src/page/Reply.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Reply = () => {
-  return;
-};
-
-export default Reply;
