--- src/page/Write.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Write = () => {
-  return;
-};
-
-export default Write;
