+++ src/pages/Home.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useSelector } from 'react-redux';
-import Header from '../components/Header';
-import TextList from '../components/TextList';
+import HomeHeader from '../components/home/HomeHeader';
+import TextList from '../components/home/TextList';
 const Home = () => {
   //   const contentsList = useSelector((state) => {
   //     return state.listReducer.list;
   //   });
   return (
     <div>
-      <Header />
+      <HomeHeader />
       <TextList />
       <TextList />
     </div>
