+++ src/pages/Write.jsx
@@ -1,18 +1,17 @@
-import Header from "../components/writeComponents/Header"
-import Form from "../components/writeComponents/Form"
-import styled from "styled-components"
+import Header from '../components/write/Header';
+import Form from '../components/write/Form';
+import styled from 'styled-components';
 
 const Write = () => {
-
   return (
     <WriteContainer>
       <Header />
       <Form />
     </WriteContainer>
-  )
-}
+  );
+};
 
-export default Write
+export default Write;
 
 const WriteContainer = styled.div`
   display: flex;
